ohja

+++ src/Modules/Members/MemberModel.php
@@ -46,7 +46,7 @@ class MemberModel
             return false;
         }
         $member = self::getMember($id);
-        if ($member->jaarlidmaatschap === $targetYear) {
+        if (($member === null) || ($member->jaarlidmaatschap === $targetYear)) {
             return false;
         }
         $member->id = null;
@@ -54,29 +54,23 @@ class MemberModel
         $member->paymentDetails->status = 0;
         $member->creationDate = null;
         $member->modificationDate = null;
-
         return (MemberMapper::new($member));
     }
 
-    /**
-     * @param int|null $id
-     */
     public static function getMember(int $id = null): ?Member
     {
-        if ($id === null) {
-            return null;
+        if ($id !== null) {
+            $membermap = MemberMapper::getMemberById($id);
+            if ($membermap !== null) {
+                return new Member($id, $membermap->jaarlidmaatschap, $membermap->voorletters, $membermap->voornaam, $membermap->achternaam, $membermap->geboortedatum, new MemberAddress($membermap->adres, $membermap->postcode, $membermap->huisnummer, $membermap->woonplaats), new MemberContactDetails($membermap->telefoon_vast, $membermap->telefoon_mobiel, $membermap->emailadres), $membermap->ingangsdatum, $membermap->geslacht, new MemberPreferences($membermap->nieuwsbrief, $membermap->vrijwilliger, $membermap->vrijwilligeroptie1, $membermap->vrijwilligeroptie2, $membermap->vrijwilligeroptie3, $membermap->vrijwilligeroptie4, $membermap->vrijwilligeroptie5), new MemberPaymentDetails($membermap->betalingswijze, $membermap->iban, $membermap->machtigingskenmerk, $membermap->status));
+            }
         }
-        $membermap = MemberMapper::getMemberById($id);
-        return new Member($id, $membermap->jaarlidmaatschap, $membermap->voorletters, $membermap->voornaam, $membermap->achternaam, $membermap->geboortedatum, new MemberAddress($membermap->adres, $membermap->postcode, $membermap->huisnummer, $membermap->woonplaats), new MemberContactDetails($membermap->telefoon_vast, $membermap->telefoon_mobiel, $membermap->emailadres), $membermap->ingangsdatum, $membermap->geslacht, new MemberPreferences($membermap->nieuwsbrief, $membermap->vrijwilliger, $membermap->vrijwilligeroptie1, $membermap->vrijwilligeroptie2, $membermap->vrijwilligeroptie3, $membermap->vrijwilligeroptie4, $membermap->vrijwilligeroptie5), new MemberPaymentDetails($membermap->betalingswijze, $membermap->iban, $membermap->machtigingskenmerk, $membermap->status));
+        return null;
     }
 
-    /**
-     * @param int $id
-     * @return bool
-     */
     public static function delete(int $id = null): bool
     {
-        if (MemberMapper::doesMemberIdExist($id)) {
+        if (($id !== null) && MemberMapper::doesMemberIdExist($id)) {
             if (MemberMapper::delete($id)) {
                 Session::add('feedback_positive', 'Lid verwijderd (ID: '.$id.').');
                 return true;
@@ -90,27 +84,24 @@ class MemberModel
 
     public static function updateMember(Member $member = null): bool
     {
-        if ($member === null) {
-            return false;
+        if (($member !== null) && MemberMapper::updateMember($member)) {
+            Session::add('feedback_positive', 'Lid opgeslagen.');
+            return true;
         }
-        return (MemberMapper::updateMember($member));
+        Session::add('feedback_positive', 'Lid niet opgeslagen.');
+        return false;
     }
 
-    /**
-     * @param Member|null $member
-     */
-    public static function createMember(Member $member = null)
+    public static function createMember(Member $member = null): bool
     {
-        if (MemberMapper::doesEmailforYearExist($member->jaarlidmaatschap, $member->contactDetails->emailadres)) {
-            Session::add('feedback_negative', 'Emailadres wordt dit jaar al gebruikt door een ander lid.');
-            Redirect::to('membership/');
-        } else {
-            if (MemberMapper::new($member)) {
-                Session::add('feedback_positive', 'Lid toegevoegd.');
-                Redirect::to('membership/');
-            }
+        if (($member === null)) {
             Session::add('feedback_negative', 'Lid toevoegen mislukt.');
-            Redirect::to('membership/');
+        } elseif (($member->jaarlidmaatschap !== null) && ($member->contactDetails->emailadres !== null) && MemberMapper::doesEmailforYearExist($member->jaarlidmaatschap, $member->contactDetails->emailadres)) {
+            Session::add('feedback_negative', 'Emailadres wordt dit jaar al gebruikt door een ander lid.');
+        } elseif (MemberMapper::new($member)) {
+            Session::add('feedback_positive', 'Lid toegevoegd.');
+            return true;
         }
+        return false;
     }
 }
